requirements.txt
discord.py
nest-asyncio
psycopg2
python-dotenv